Adding convert and open commands that use citeproc
It might make sense to change the localleader mappings too. Not sure:
processing with citeproc takes more time. But I for one always want
it when converting to PDF or ODT, if not HTML.

Or we could add mappings for both, but I want to be minimalist with
mappings (maybe that is wrongheaded).