5e4c4c6 broke bibfile detection. This fixes it.
The Pandoc_Find_Bibfile function actually works in reverse: it finds
every possible match, with the most preferred being the last one it
finds. The change made in 5e4c4c6 meant that it would set g:pandoc_bibfile
to the empty string if it failed to find its most preferred bibfile.

Now it starts by setting g:pandoc_bibfile to the empty string, and then
looks for something better.